Decimal value was done

+++ app/budgeting/models.py
@@ -2,6 +2,7 @@
 from __future__ import unicode_literals
 
 from django.db import models
+from decimal import getcontext
 
 class Budgeting(models.Model):
     TYPE_PROPOSAL_CHOICES = (
@@ -28,7 +29,7 @@ class Budgeting(models.Model):
 class Item(models.Model):
     coa = models.CharField(max_length=100, null=True, blank=True)
     description = models.CharField(max_length=220)
-    qty = models.DecimalField(max_digits=11, decimal_places=2, default=0)
+    qty = models.IntegerField(default=0)
     amount = models.DecimalField(max_digits=20, decimal_places=2, default=0)
     tot_amount = models.DecimalField(max_digits=20, decimal_places=2, default=0)
     budgeting = models.ForeignKey(Budgeting, on_delete=models.CASCADE)
